requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
streamlit
python-dotenv
fastapi
uvicorn
